+++ R/utilities.R
@@ -191,7 +191,7 @@ samplekan <- function(set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"), siz
 # curves in the d argument)
 # The main work is done with code from the non-CRAN package svgparser v0.1.2 (MIT license, see svgparser_lite.R)
 # Code had to be extracted because it seems otherwise kanjistat cannot be on CRAN
-.kanjivg_to_list <- function(xml, padhex, char, bezier_discr = c("svgparser", "eqtimed", "eqspaced"),
+.kanjivg_to_list <- function(xml, padhex, char, bezier_discr = c("svgparser", "eqtimed", "eqspaced"), code=1,
                              flatten_inner = TRUE, flatten_leaves = TRUE) {  # padhex and char are for verification purposes
   # check if everything is ok with xml
   # (it is not impossible that some of the kanjivg-files do not pass these tests!)
@@ -262,12 +262,19 @@ samplekan <- function(set = c("kyouiku", "jouyou", "jinmeiyou", "kanjidic"), siz
         y <- 1-points_df$y/109
         li <- cbind(x[-1],y[-1])    # the first coordinates are from the move instruction, hence the same as the second coords.
       } else {
-        li <- points_from_svg(path_d, 50/109, eqspaced=eqspaced)
+        if (code == 1) {
+          li <- points_from_svg(path_d, 50/109, eqspaced=eqspaced)
+        } else {
+          li <- points_from_svg2(path_d, 50/109, eqspaced=eqspaced)
+        }
         li <- rescale_points(li, a=c(1,-1)/109, b=c(0,1))
       }
       attr(li, "id") <- id
       attr(li, "type") <- type
-      attr(li, "d") <- path_d # we save the original Bézier curve as attribute
+      attr(li, "d") <- path_d # we save the original d-string describing the Bézier curves of the stroke as attribute
+      attr(li, "beziermat") <- strictformat_bezier(path_d)
+         # as well as our 2 x (1+3n) matrix format specifying MCC..C points without duplication
+         # (where n denotes the number of curves in the stroke) 
     }
     li
   }